# server/docs/swagger.yaml
basePath: /
definitions:
  apidocs.AuditLogsListResponse:
    properties:
      audit_logs:
        items:
          $ref: '#/definitions/models.AuditLog'
        type: array
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
...
  apidocs.CreatePaymentAccountRequest:
    properties:
      account_name:
        example: Tofex Ltd
        type: string
      account_number:
        example: "0123456789"
        type: string
      active:
        example: true
        type: boolean
      bank_name:
        example: Example Bank
        type: string
      branch:
        example: Lagos Main
        type: string
      currency:
        example: NGN
        type: string
      display_label:
        example: Main account
        type: string
    type: object
...
  apidocs.ErrorBody:
    properties:
      code:
        example: validation_error
        type: string
      message:
        example: invalid request
        type: string
    type: object
  apidocs.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/apidocs.ErrorBody'
    type: object
...
  apidocs.OrdersListResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/models.Order'
        type: array
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
  apidocs.PatchOrderRequest:
    properties:
      status:
        enum:
        - pending
        - confirmed
        - ready
        - completed
        - cancelled
        example: confirmed
        type: string
    type: object
  apidocs.PatchPaymentAccountRequest:
    properties:
      account_name:
        type: string
      account_number:
        type: string
      active:
        type: boolean
      bank_name:
        type: string
      branch:
        type: string
      currency:
        type: string
      display_label:
        type: string
    type: object
...
  apidocs.PaymentAccountsListResponse:
    properties:
      payment_accounts:
        items:
          $ref: '#/definitions/models.PaymentAccount'
        type: array
    type: object
...
  models.AuditLog:
    properties:
      action:
        type: string
      created_at:
        type: string
      id:
        type: integer
      ip:
        type: string
      metadata:
        type: string
      resource_id:
        type: string
      resource_type:
        type: string
      user_agent:
        type: string
      user_id:
        type: integer
    type: object
  models.Order:
    properties:
      created_at:
        type: string
      customer_email:
        type: string
      customer_name:
        type: string
      customer_note:
        type: string
      customer_phone:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      payment_account_id:
        type: integer
      payment_snapshot:
        type: string
      receipts:
        items:
          $ref: '#/definitions/models.OrderReceipt'
        type: array
      status:
        type: string
      total_cents:
        type: integer
      updated_at:
        type: string
    type: object
  models.OrderItem:
    properties:
      id:
        type: integer
      order_id:
        type: string
      product_id:
        type: integer
      product_name:
        type: string
      quantity:
        type: integer
      unit_price_cents:
        type: integer
    type: object
  models.OrderReceipt:
    properties:
      created_at:
        type: string
      id:
        type: integer
      mime:
        type: string
      order_id:
        type: string
      path:
        type: string
      size_bytes:
        type: integer
    type: object
  models.PaymentAccount:
    properties:
      account_name:
        type: string
      account_number:
        type: string
      active:
        type: boolean
      bank_name:
        type: string
      branch:
        type: string
      created_at:
        type: string
      currency:
        type: string
      display_label:
        type: string
      id:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for Tofex restaurant/confectionery chain (staff admin + public
    guest ordering).
  title: Tofex API
  version: "1.0"
paths:
  /api/v1/admin/audit-logs:
    get:
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 50
        description: Per page
        in: query
        name: per_page
        type: integer
      - description: Filter by staff user ID
        in: query
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.AuditLogsListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List audit logs
      tags:
      - admin-audit
...
  /api/v1/admin/orders:
    get:
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Per page
        in: query
        name: per_page
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.OrdersListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - admin-orders
  /api/v1/admin/orders/{id}:
    get:
      parameters:
      - description: Order UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get order
      tags:
      - admin-orders
    patch:
      consumes:
      - application/json
      parameters:
      - description: Order UUID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.PatchOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update order status
      tags:
      - admin-orders
  /api/v1/admin/orders/{id}/receipts/{rid}/file:
    get:
      parameters:
      - description: Order UUID
        in: path
        name: id
        required: true
        type: string
      - description: Receipt ID
        in: path
        name: rid
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download order receipt
      tags:
      - admin-orders
  /api/v1/admin/payment-accounts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.PaymentAccountsListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List payment accounts (admin)
      tags:
      - admin-payment-accounts
    post:
      consumes:
      - application/json
      parameters:
      - description: Bank account
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.CreatePaymentAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.PaymentAccount'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create payment account
      tags:
      - admin-payment-accounts
  /api/v1/admin/payment-accounts/{id}:
    delete:
      parameters:
      - description: Payment account ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: no content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete payment account
      tags:
      - admin-payment-accounts
    patch:
      consumes:
      - application/json
      parameters:
      - description: Payment account ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.PatchPaymentAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentAccount'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update payment account
      tags:
      - admin-payment-accounts
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"